+++ cvat-ui/src/reducers/notifications-reducer.ts
@@ -99,6 +99,9 @@ const defaultState: NotificationsState = {
         models: {
             inferenceDone: '',
         },
+        auth: {
+            changePasswordDone: '',
+        }
     },
 };
